refactor(svgUtils): decompose 1807-line file into focused modules
The largest frontend source file, flagged by the continuous quality
reporter. Split into 8 modules under `frontend/src/utils/svg/`, each
with a single responsibility. `svgUtils.ts` is now a 60-line barrel
that re-exports every symbol — every caller keeps working unchanged.

Module              Lines   Responsibility
─────────────────── ─────── ─────────────────────────────────────────
idMap.ts              29    Cached DOM-id map for an SVG container
svgBoost.ts          122    Dynamic font/node scaling + processSvg
metadataIndex.ts      40    Build MetadataIndex from pypowsybl meta
highlights.ts        422    Overloaded / action-target / contingency
                            halos; line & VL target resolution
deltaVisuals.ts      137    Delta flow coloring + text replacement
actionPinData.ts     344    Pin descriptors + severity palette +
                            anchor resolution (pure, no DOM)
actionPinRender.ts   537    DOM injection for pins + highlights +
                            click semantics + scale math
fitRect.ts           125    Padded viewBox computation

Largest frontend file is now App.tsx at 1370 lines — the state
orchestration hub, exempt from the component-size ceiling by design.

New helpers promoted to the public API (formerly inline closures):

- formatPinLabel(details)       — percentage / DIV / ISL / em-dash
- formatPinTitle(idLabel, det)  — hover tooltip string
- fanOutColocatedPins(pins)     — circular fan-out of colocated pins
- curveMidpoint(p1, p2)         — quadratic Bezier midpoint + ctrl
- computePinScale(...)          — pin scale math for rescaler
- rectFromBounds (internal)     — shared MIN_SPAN + padding rules

Coverage
- 61 new unit tests across 5 co-located test files (idMap,
  svgBoost, metadataIndex, actionPinData, actionPinRender,
  fitRect).
- Existing svgUtils.test.ts (144 tests) unchanged and all green.
- Full frontend suite: 1000 tests passing.
- TypeScript strict build clean; ESLint zero warnings.

Docs
- Added §11 Delta in docs/architecture/code-quality-analysis.md
  with the module table and test coverage list.
