+++ frontend/src/utils/svg/metadataIndex.ts
@@ -1,0 +1,40 @@
+// Copyright (c) 2025-2026, RTE (https://www.rte-france.com)
+// This Source Code Form is subject to the terms of the Mozilla Public License, version 2.0.
+// If a copy of the Mozilla Public License, version 2.0 was not distributed with this file,
+// you can obtain one at http://mozilla.org/MPL/2.0/.
+// SPDX-License-Identifier: MPL-2.0
+
+import type { MetadataIndex, NodeMeta, EdgeMeta } from '../../types';
+
+/**
+ * Build Map indices from pypowsybl metadata for O(1) lookups.
+ *
+ * Accepts either a raw JSON string or a pre-parsed object. Returns
+ * null when the input is missing.
+ */
+export const buildMetadataIndex = (metadata: unknown): MetadataIndex | null => {
+    if (!metadata) return null;
+    const meta = typeof metadata === 'string' ? JSON.parse(metadata) : metadata;
+    const nodes: NodeMeta[] = (meta as { nodes?: NodeMeta[] }).nodes || [];
+    const edges: EdgeMeta[] = (meta as { edges?: EdgeMeta[] }).edges || [];
+
+    const nodesByEquipmentId = new Map<string, NodeMeta>();
+    const nodesBySvgId = new Map<string, NodeMeta>();
+    const edgesByEquipmentId = new Map<string, EdgeMeta>();
+    const edgesByNode = new Map<string, EdgeMeta[]>();
+
+    nodes.forEach(n => {
+        nodesByEquipmentId.set(n.equipmentId, n);
+        nodesBySvgId.set(n.svgId, n);
+    });
+
+    edges.forEach(e => {
+        edgesByEquipmentId.set(e.equipmentId, e);
+        if (!edgesByNode.has(e.node1)) edgesByNode.set(e.node1, []);
+        edgesByNode.get(e.node1)!.push(e);
+        if (!edgesByNode.has(e.node2)) edgesByNode.set(e.node2, []);
+        edgesByNode.get(e.node2)!.push(e);
+    });
+
+    return { nodesByEquipmentId, nodesBySvgId, edgesByEquipmentId, edgesByNode };
+};
